+++ Cargo.toml
@@ -40,8 +40,7 @@ chrono = { version = "0.4", features = ["serde"] }
 eyre = "0.6"
 directories = "4"
 indicatif = "0.16.2"
-serde_derive = "1.0.125"
-serde = "1.0.126"
+serde = { version = "1.0.126", features = ["derive"] }
 serde_json = "1.0.75"
 tui = "0.16"
 termion = "1.5"
