+++ atuin-common/src/api.rs
@@ -1,8 +1,5 @@
-use std::borrow::Cow;
-
-use axum::{response::IntoResponse, Json};
 use chrono::Utc;
-use serde::Serialize;
+use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct UserResponse {
@@ -55,35 +52,4 @@ pub struct SyncHistoryRequest {
 #[derive(Debug, Serialize, Deserialize)]
 pub struct SyncHistoryResponse {
     pub history: Vec<String>,
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-pub struct ErrorResponse<'a> {
-    pub reason: Cow<'a, str>,
-}
-
-impl<'a> IntoResponse for ErrorResponseStatus<'a> {
-    fn into_response(self) -> axum::response::Response {
-        (self.status, Json(self.error)).into_response()
-    }
-}
-
-pub struct ErrorResponseStatus<'a> {
-    pub error: ErrorResponse<'a>,
-    pub status: http::StatusCode,
-}
-
-impl<'a> ErrorResponse<'a> {
-    pub fn with_status(self, status: http::StatusCode) -> ErrorResponseStatus<'a> {
-        ErrorResponseStatus {
-            error: self,
-            status,
-        }
-    }
-
-    pub fn reply(reason: &'a str) -> ErrorResponse {
-        Self {
-            reason: reason.into(),
-        }
-    }
 }
