+++ atuin-common/src/lib.rs
@@ -1,7 +1,4 @@
 #![forbid(unsafe_code)]
 
-#[macro_use]
-extern crate serde_derive;
-
 pub mod api;
 pub mod utils;
