+++ atuin-server/Cargo.toml
@@ -17,8 +17,7 @@ eyre = "0.6"
 uuid = { version = "0.8", features = ["v4"] }
 whoami = "1.1.2"
 config = "0.13"
-serde_derive = "1.0.125"
-serde = "1.0.126"
+serde = { version = "1.0.126", features = ["derive"] }
 serde_json = "1.0.75"
 sodiumoxide = "0.2.6"
 base64 = "0.13.0"
