+++ atuin-server/src/lib.rs
@@ -8,9 +8,6 @@ use eyre::{Context, Result};
 
 use crate::settings::Settings;
 
-#[macro_use]
-extern crate serde_derive;
-
 pub mod auth;
 pub mod calendar;
 pub mod database;
