+++ atuin-server/src/settings.rs
@@ -1,4 +1,5 @@
 use fs_err::{create_dir_all, File};
+use serde::{Deserialize, Serialize};
 use std::io::prelude::*;
 use std::path::PathBuf;
